HW 5 - Producing Data
3 parts to today's homework:

1. Using R to determine the Mean and Standard Deviation, number of subjects in the 10 samples, and summarizing the above through a small discussion of the data.

2. Taking an abstract from a predetermined article and answering 5 questions about the data and information provided. Most of the necessary info has been provided so not much needs to be done in R, this is more so of an analytical, hands on approach.

3. More in-depth practice with R. Today we examine a sample of 100 volunteers, 30 whom are smokers and 70 non-smokers. 4 parts to this:

Part 1  where we examine a random selection of 10 smokers and 10 non-smokers and display the label in ascending order. The random sample is a single satrified sample.

Part 2 has us repeating the samples and performing 12 samples of 10 smokers from the population of 30, and for each smoker we record the number of samples in which the smoker appears in, reporting the number at the end.

Part 3 involves us getting the mean and standard deviation of the dataset. We also need to identify why the mean is exactly 4.

Part 4 plots a histogram of all our data we've collected, specifically showing the number of times someone has appeared: 0, 1, 2, ... etc.

Signed-off-by: CieranW <cieranwong@gmail.com>