Init Commit for HW 6: Basics of Probability
Question 3 focuses on playing with probability in R. We've got a large sample size of tortoises, 300 to be exact. 50 of those 300 are tagged and released. After a certain amount of time had passed, 15 of the 300 are caught again.

Part 1 looks at the probability of one of the 15 captured tortoises being one of the 50 tagged.

Part 2 asks us to use R to find the probability of each tortoise being tagged, ex. 0, 1, 2, ... 15.

Part 3 has us perform a simulation of 1000 captures, tags, and releases with a table at the end to format the outcome.

Signed-off-by: CieranW <cieranwong@gmail.com>