Fleshed out Q3
Finished writing Q3 for the hw.

Part 1 has some variables that are universal so might shift that up a bit so it's a bit easier to read. The rest of it is a probability thing where we find the probability of 1 of the 15 being tagged. Utilized dhyper here to bring in the population and the various tagged, 2nd capture, etc.

Part 2 has a for loop to go through the probabilities of the tagged count ranging from 0 to 15. Used data.frame to get it in a nice table, rounded to 4 decimal places. Sum of the probabilities should equal 1.

Part 3 we ran simulations through the use of another for loop and sample() in the for loop. Within the for loop we also do a count to find the number of captured tortoises that were tagged. Again outputted to a table using data.frame with some headers/labels.

Signed-off-by: CieranW <cieranwong@gmail.com>